Phase 5 (step 1): units of measure + decimal quantities
Foundation for the projects/planning workflow: items can now be measured, not
just counted (e.g. 25.5 ft of wire), so partial amounts are representable.

- InventoryItem.Quantity int -> decimal(18,3); new nullable Unit (e.g. ft, m,
  box). Blank unit = a plain count. EF migration converts existing integer
  quantities losslessly and adds the Unit column.
- DTOs/services updated: item create/read/list, dashboard total quantity,
  reports, CSV import/export (new Unit column, decimal quantity parse), and JSON
  backup/restore all carry Unit and decimal quantity.
- UI: item form gains a Unit field (datalist of common units) and a decimal
  quantity input; item list/detail and printable reports show "25 ft" / "x3"
  via CommonUnits.Format.

Verification: builds with zero warnings; 54/54 tests pass; upgrade path tested
(legacy INTEGER Quantity=7 preserved through the migration, Unit null for old
rows); live round-trip confirmed (created 25.5 ft item reads back correctly,
dashboard total is decimal, CSV export includes Unit).

+++ src/HomeStock.Application/Models/DashboardModels.cs
@@ -5,7 +5,7 @@ public record CountByName(string Name, int Count);
 /// <summary>Aggregate figures and lists shown on the dashboard.</summary>
 public record DashboardDto(
     int TotalItemRecords,
-    int TotalQuantity,
+    decimal TotalQuantity,
     decimal EstimatedTotalValue,
     int LoanedOutCount,
     int MissingPhotoCount,
